requirements.txt
paho-mqtt
pyvesync